@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.7",
+    "@angular/animations": "^14.1.3",
     "@angular/cdk": "^13.3.9",
-    "@angular/common": "~13.2.7",
-    "@angular/compiler": "~13.2.7",
-    "@angular/core": "~13.2.7",
-    "@angular/forms": "~13.2.7",
+    "@angular/common": "^14.1.3",
+    "@angular/compiler": "^14.1.3",
+    "@angular/core": "^14.1.3",
+    "@angular/forms": "^14.1.3",
     "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.2.7",
-    "@angular/platform-browser-dynamic": "~13.2.7",
-    "@angular/router": "~13.2.7",
+    "@angular/platform-browser": "^14.1.3",
+    "@angular/platform-browser-dynamic": "^14.1.3",
+    "@angular/router": "^14.1.3",
     "@fortawesome/angular-fontawesome": "^0.10.2",
     "@fortawesome/fontawesome-svg-core": "^6.1.0",
     "@fortawesome/free-brands-svg-icons": "^6.1.0",
@@ -31,9 +31,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.9",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~13.2.7",
+    "@angular-devkit/build-angular": "^14.1.3",
+    "@angular/cli": "^14.1.3",
+    "@angular/compiler-cli": "^14.1.3",
     "@tailwindcss/aspect-ratio": "^0.4.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
@@ -46,6 +46,6 @@
     "karma-jasmine-html-reporter": "~1.7.0",
     "postcss": "^8.4.6",
     "tailwindcss": "^3.0.18",
-    "typescript": "~4.5.2"
+    "typescript": "~4.7.4"
   }
 }
